requirements.txt
babel==2.17.0
langchain==1.0.5
langchain-google-genai==3.0.1
pandas==2.3.3
pillow==12.0.0
sentencepiece==0.2.1
streamlit==1.51.0
torch==2.9.0
transformers==4.57.1
typing_extensions==4.15.0
xmltodict==1.0.2
easyocr
opencv-python
numpy